front: rebuild folder structures
