+++ frontend/src/components/Copyright.tsx
