+++ frontend/src/components/Footer.tsx
