+++ frontend/src/components/Layout.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import MainFooter from './Footer';
+import MainHeader from './Header';
+
+const CustomerLayout = (props:any) => {
+  return (
+    <div>
+      <MainHeader />
+      
+      <MainFooter />
+    </div>
+  );
+};
+
+export default CustomerLayout;
