+++ frontend/src/components/SignupForm.tsx
@@ -1,99 +1,28 @@
-import React, { useState, useEffect } from 'react';
 import Avatar from '@material-ui/core/Avatar';
 import Button from "@material-ui/core/Button";
 import TextField from '@material-ui/core/TextField';
-import CssBaseline from '@material-ui/core/CssBaseline';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Checkbox from '@material-ui/core/Checkbox';
 import Grid from '@material-ui/core/Grid';
-import Box from '@material-ui/core/Box';
 import CircularProgress from '@material-ui/core/CircularProgress';
 import AccountCircleIcon from '@material-ui/icons/AccountCircle';
 import Typography from '@material-ui/core/Typography';
-import { makeStyles } from '@material-ui/core/styles';
-import Container from '@material-ui/core/Container';
-import { connect } from 'react-redux';
-import { Redirect } from 'react-router-dom';
-import { authSignup } from '../../store/actions/cauth';
-import { FormData } from '../../types/types/auth';
-import Copyright from '../other/Copyright';
+//Custom
+import { TAuthState, SignUpFormData } from '../types/TypeAuth';
+import { AuthStyles } from './Styles';
 
-const useStyles = makeStyles((theme) => ({
-  paper: {
-    marginTop: theme.spacing(8),
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-  },
-  avatar: {
-    margin: theme.spacing(1),
-    backgroundColor: theme.palette.secondary.light,
-    width: theme.spacing(7),
-    height: theme.spacing(7),
-  },
-  form: {
-    width: '100%',
-    marginTop: theme.spacing(1),
-  },
-  submit: {
-    margin: theme.spacing(3, 0, 2),
-  },
-}));
+type SignupProps = {
+  state: TAuthState;
+  signUpFormData: SignUpFormData
+  formDataValid: SignUpFormData;
+  handleChange: (event: React.ChangeEvent<HTMLInputElement>) => void;
+  handleSignup: (event: React.FormEvent<HTMLFormElement>) => void;
+}
 
-const Signup = (props: any) => {
-  const classes = useStyles();
 
-  const [values, setValues] = useState({
-    username: '',
-    holyname: '',
-    name: '',
-    email: '',
-    address: '',
-    password1: '',
-    password2: ''
-  });
-
-  const [errors, setErrors] = useState({} as FormData);
-  const { loading, token } = props;
-
-  const formDataValid: FormData = {
-    username: '',
-    holyname: '',
-    name: '',
-    email: '',
-    address: '',
-    password1: '',
-    password2: '',
-    message: ''
-  }
-
-  const handleSignup = (e: any) => {
-    e.preventDefault();
-    if (values.username.length <= 8) {         // client side validation here
-      formDataValid.username = 'Tên đăng nhập phải dài hơn 8 ký tự';
-      setErrors(formDataValid);
-    } else {
-      formDataValid.username = '';
-      setErrors(formDataValid);
-      const { username, email, password1, password2 } = values;
-      props.signup(username, email, password1, password2);
-    }
-  }
-
-  const handleChange = (e: any) => {
-    setValues((values) => ({
-      ...values, [e.target.name]: e.target.value,
-    }))
-  }
-
-  if (token) {
-    return (
-      <Redirect to='/' /> //welcome popup show here
-    )
-  } else {
-    return (
-      <Container component='main' maxWidth='xs'>
-        <CssBaseline />
+const SignupForm : React.FC<SignupProps> = ({ state, signUpFormData,formDataValid, handleChange, handleSignup }) => {
+  const classes = AuthStyles();
+  return (
         <div className={classes.paper}>
           <Avatar className={classes.avatar}>
             <AccountCircleIcon fontSize="large" />
@@ -111,14 +40,14 @@ const Signup = (props: any) => {
                   autoComplete='username'
                   name='username'
                   variant='outlined'
-                  value={values.username}
+                  value={signUpFormData.username}
                   required
                   fullWidth
                   id='username'
                   label='Tên đăng nhập'
                   onChange={handleChange}
-                  helperText={errors.username}
-                  error={!!errors.username}
+                  helperText={formDataValid.username}
+                  error={!!formDataValid.username}
                   autoFocus
                 />
               </Grid>
@@ -127,14 +56,14 @@ const Signup = (props: any) => {
                   autoComplete='holyname'
                   name='holyname'
                   variant='outlined'
-                  value={values.holyname}
+                  value={signUpFormData.holyname}
                   required
                   fullWidth
                   id='holyname'
                   label='Tên Thánh'
                   onChange={handleChange}
-                  helperText={errors.holyname}
-                  error={!!errors.holyname}
+                  helperText={formDataValid.holyname}
+                  error={!!formDataValid.holyname}
                 />
               </Grid>
               <Grid item xs={12} sm={6}>
@@ -142,14 +71,14 @@ const Signup = (props: any) => {
                   autoComplete='name'
                   name='name'
                   variant='outlined'
-                  value={values.name}
+                  value={signUpFormData.name}
                   required
                   fullWidth
                   id='name'
                   label='Họ và Tên'
                   onChange={handleChange}
-                  helperText={errors.name}
-                  error={!!errors.name}
+                  helperText={formDataValid.name}
+                  error={!!formDataValid.name}
                 />
               </Grid>
               <Grid item xs={12}>
@@ -157,13 +86,13 @@ const Signup = (props: any) => {
                   autoComplete='email'
                   name='email'
                   variant='outlined'
-                  value={values.email}
+                  value={signUpFormData.email}
                   required
                   fullWidth
                   id='email'
                   label='Địa chỉ email'
-                  helperText={errors.email}
-                  error={!!errors.email}
+                  helperText={formDataValid.email}
+                  error={!!formDataValid.email}
                   onChange={handleChange}
                 />
               </Grid>
@@ -172,13 +101,13 @@ const Signup = (props: any) => {
                   autoComplete='address'
                   name='address'
                   variant='outlined'
-                  value={values.address}
+                  value={signUpFormData.address}
                   required
                   fullWidth
                   id='address'
                   label='Địa chỉ chỗ ở hiện tại'
-                  helperText={errors.address}
-                  error={!!errors.address}
+                  helperText={formDataValid.address}
+                  error={!!formDataValid.address}
                   onChange={handleChange}
                 />
               </Grid>
@@ -186,7 +115,7 @@ const Signup = (props: any) => {
                 <TextField
                   autoComplete='password1'
                   name='password1'
-                  value={values.password1}
+                  value={signUpFormData.password1}
                   onChange={handleChange}
                   variant='outlined'
                   required
@@ -194,23 +123,23 @@ const Signup = (props: any) => {
                   id='password1'
                   label='Mật khẩu'
                   type="password"
-                  helperText={errors.password1}
-                  error={!!errors.password1}
+                  helperText={formDataValid.password1}
+                  error={!!formDataValid.password1}
                 />
               </Grid>
               <Grid item xs={12}>
                 <TextField
                   autoComplete='password2'
                   name='password2'
-                  value={values.password2}
+                  value={signUpFormData.password2}
                   onChange={handleChange}
                   variant='outlined'
                   required
                   fullWidth
                   id='password2'
                   label='Nhập lại mật khẩu'
-                  helperText={errors.password2}
-                  error={!!errors.password2}
+                  helperText={formDataValid.password2}
+                  error={!!formDataValid.password2}
                   type="password"
                 />
               </Grid>
@@ -226,38 +155,17 @@ const Signup = (props: any) => {
               variant='contained'
               fullWidth
               color='primary'
-              disabled={loading}
+              disabled={state.loading}
               className={classes.submit}
             >
               Đăng ký
-              {loading && (
+              {state.loading && (
                 <CircularProgress size={30} color="secondary" />
               )}
             </Button>
           </form>
         </div>
-        <Box mt={5}>
-          <Copyright />
-        </Box>
-      </Container>
     );
-  }
 };
 
-//Map the states to local props
-const mapStateToProps = (state: any) => ({
-  loading: state.auth.loading,
-  error: state.auth.error,
-  token: state.auth.token,
-});
-
-// map actions to local props in this functional component
-
-const mapActionsToProps = (dispatch: any) => {
-  return {
-    signup: (username: string, email: string, password1: string, password2: string) =>
-      dispatch(authSignup(username, email, password1, password2)),
-  };
-};
-
-export default connect(mapStateToProps, mapActionsToProps)(Signup);
+export default SignupForm;
