+++ frontend/src/components/main.css
