+++ frontend/src/components/newfeed.css
