--- frontend/src/store/utility.ts
@@ -1,6 +1,0 @@
-export const updateObject = (oldObject: any, updateProperties: any) => {
-  return {
-    ...oldObject,
-    ...updateProperties,
-  };
-};
