+++ frontend/src/types/IUser.ts
@@ -1,0 +1,8 @@
+export interface IUser {
+  id: number;
+  username: string;
+  holyname:string;
+  name:string;
+  email: string;
+  address:string;
+}
