--- frontend/src/types/interfaces/acticel.ts
@@ -1,12 +1,0 @@
-export interface Acticle {
-  id: number;
-  acticle_title: string;
-  acticle_writer: string;
-  acticle_date: string;
-  acticle_bref: string;
-  acticle_content: string;
-  acticle_liked: number;
-  acticle_shared: number;
-  acticle_image: string;
-  acticle_category: string;
-}
