--- frontend/src/types/interfaces/newfeed.ts
@@ -1,13 +1,0 @@
-export interface INewFeed {
-  id: number;
-  nf_title: string;
-  nf_type: string;
-  nf_image: string;
-  nf_brief_content: string;
-  nf_content: string;
-  nf_language: string;
-  nf_date_created: string;
-  nf_post_like: number;
-  nf_post_share: number;
-  nf_post_clicked: number;
-}
