--- frontend/src/types/interfaces/province.ts
@@ -1,6 +1,0 @@
-export interface IProvince {
-  id: number;
-  province_nation: string;
-  province_code: string;
-  province_en_name: string;
-}
