--- frontend/src/types/interfaces/register.ts
@@ -1,5 +1,0 @@
-export interface IRegister {
-  id: number;
-  mass_id: number;
-  user_status: string;
-}
